Add collectible candy system to player and scene
Introduces a Candy.gd script and collectible candy objects to the MechanicsTestScene. Updates the player node and script to detect and collect candy, incrementing a counter and updating the UI label. Also adds a CollisionSensor to the player for candy detection and configures relevant input and group settings in project.godot.